requirements.txt
docker_py==1.10.6
Flask==3.1.0
setuptools==75.8.0
